Changed data types to clear warnings at /W4 (mex level) on MSVC. Also cleared the MD5_ functions to use standard memset/memcpy. While at it, removed the type defines and replaces the proper instances.
Clearing warnings is fun, because you find bugs like the one on TextFile.cpp CTextFile::WriteLine. Notice how wxEmptyString has a implicit conversion to wxChar* that will always be bool "true", meaning we return that we wrote the line correctly when it failed because the function returns bool, not wxString.

+++ trunk/src/libs/common/Format.cpp
@@ -185,7 +185,7 @@ void CFormat::SetCurrentField(const wxString& value)
 
 	// Format strings are expected to follow the folllowing structure:
 	// 	%[Flags][Width][.Precision][Length]<Type>
-	for (unsigned i = m_fieldStart + m_fieldLength; i < m_format.Length(); ++i) {
+	for (size_t i = m_fieldStart + m_fieldLength; i < m_format.Length(); ++i) {
 		const wxChar c = m_format[i];
 
 		if (pos >= PosStart) {
@@ -216,7 +216,7 @@ void CFormat::SetCurrentField(const wxString& value)
 				break;
 			}
 		} else if (c == wxT('%')) {
-			const unsigned offset = m_fieldStart + m_fieldLength;
+			const size_t offset = m_fieldStart + m_fieldLength;
 			// If there was anything before this, then prepend it.
 			if (offset < i) {
 				m_result += m_format.Mid(offset, i - offset);
